app/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.7.0
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
fastapi==0.115.6
flower==2.0.1
h11==0.14.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
humanize==4.11.0
idna==3.10
iniconfig==2.0.0
kombu==5.4.2
packaging==24.2
pluggy==1.5.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psycopg2==2.9.10
pydantic==2.10.4
pydantic_core==2.27.2
pytest==8.3.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-status==1.0.1
pytz==2024.2
PyYAML==6.0.2
redis==5.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
uvicorn==0.34.0
uvloop==0.21.0
vine==5.1.0
watchfiles==1.0.3
wcwidth==0.2.13
websockets==14.1
